# tickets/public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ticket Sales API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/events/{event}/ticket-types": {
            "get": {
                "tags": [
                    "Ticket Types"
                ],
                "summary": "Get ticket types for an event",
                "description": "Returns paginated list of ticket types with optional filters",
                "operationId": "998281a1a61c1b697f15eda0a48b2b81",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "description": "Filter by active status",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Minimum price",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Maximum price",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort field",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "price",
                                "name",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "description": "Sort direction",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of ticket types"
                    },
                    "404": {
                        "description": "No ticket types found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ticket Types"
                ],
                "summary": "Create ticket type",
                "description": "Admin only - Creates a new ticket type for an event",
                "operationId": "11fc870f0725b0eeb0caaf9db1c2dd66",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "quantity_total"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "VIP"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Premium"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 5000
                                    },
                                    "quantity_total": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "quantity_sold": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "sales_start_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-06-01 10:00:00"
                                    },
                                    "sales_end_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-07-15 18:00:00"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket type created successfully"
                    },
                    "403": {
                        "description": "Only admins can create ticket types"
                    }
                }
            }
        },
        "/api/ticket-types/{ticketType}": {
            "get": {
                "tags": [
                    "Ticket Types"
                ],
                "summary": "Get single ticket type",
                "description": "Returns ticket type details with event",
                "operationId": "ffbf717fc37111e97d7dd941f1d28b7f",
                "parameters": [
                    {
                        "name": "ticketType",
                        "in": "path",
                        "description": "Ticket Type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket type details"
                    },
                    "404": {
                        "description": "Ticket type not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Ticket Types"
                ],
                "summary": "Update ticket type",
                "description": "Admin only - Updates ticket type details",
                "operationId": "c81d7706b9abe5413cc0cef6e8e1af1d",
                "parameters": [
                    {
                        "name": "ticketType",
                        "in": "path",
                        "description": "Ticket Type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "quantity_total": {
                                        "type": "integer"
                                    },
                                    "sales_start_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "sales_end_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ticket type updated successfully"
                    },
                    "403": {
                        "description": "Only admins can update ticket types"
                    },
                    "404": {
                        "description": "Ticket type not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ticket Types"
                ],
                "summary": "Delete ticket type",
                "description": "Admin only - Deletes a ticket type",
                "operationId": "8a959a2b5655b6123e28ddf213f90ebf",
                "parameters": [
                    {
                        "name": "ticketType",
                        "in": "path",
                        "description": "Ticket Type ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket type deleted successfully"
                    },
                    "403": {
                        "description": "Only admins can delete ticket types"
                    },
                    "404": {
                        "description": "Ticket type not found"
                    }
                }
            }
        }
    },
...
}